# docs/openapi.json
{"openapi":"3.0.1","info":{"title":"Microservicio de Plazoleta - PowerUp Plazoleta","description":"API REST para la gestión de restaurantes, menús, platos y ciclo de vida de los pedidos en la plataforma de Plazoleta de Comidas bajo Arquitectura Hexagonal.","termsOfService":"http://swagger.io/terms/","contact":{"name":"Valentina Pinto","email":"valentina.pinto@pragma.com.co"},"license":{"name":"Apache 2.0","url":"http://springdoc.org"},"version":"1.0.0"},"servers":[{"url":"http://localhost:8082","description":"Generated server url"}],"tags":[{"name":"Platos","description":"Endpoints para la creación, configuración y gestión de platos"},{"name":"Restaurantes","description":"Endpoints para la creación, configuración y gestión de restaurantes"}],"paths":{"/api/v1/dish/{id}":{"put":{"tags":["Platos"],"summary":"Modificar el precio y la descripción de un plato existente","operationId":"actualizarPlato","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateDishRequestDto"}}},"required":true},"responses":{"400":{"description":"Datos de entrada inválidos o precio menor/igual a cero"},"404":{"description":"Plato no encontrado"},"200":{"description":"Plato modificado exitosamente"}}}},"/api/v1/restaurant":{"post":{"tags":["Restaurantes"],"summary":"Crear un nuevo restaurante","operationId":"guardarRestaurante","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RestaurantRequestDto"}}},"required":true},"responses":{"400":{"description":"Datos de entrada inválidos o reglas de negocio no cumplidas"},"409":{"description":"El NIT del restaurante ya se encuentra registrado"},"201":{"description":"Restaurante creado exitosamente"}}}},"/api/v1/dish":{"post":{"tags":["Platos"],"summary":"Crear un nuevo plato asociado a un restaurante","operationId":"guardarPlato","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DishRequestDto"}}},"required":true},"responses":{"201":{"description":"Plato creado exitosamente"},"404":{"description":"Restaurante o categoría no encontrados"},"400":{"description":"Datos de entrada inválidos o precio menor/igual a cero"}}}}},"components":{"schemas":{"UpdateDishRequestDto":{"required":["descripcion","precio"],"type":"object","properties":{"precio":{"minimum":1,"type":"integer","description":"Nuevo precio del plato (número entero positivo mayor a 0)","format":"int32","example":38000},"descripcion":{"type":"string","description":"Nueva descripción detallada del plato","example":"Deliciosa bandeja paisa tradicional con ingredientes frescos y porción extra de aguacate"}},"description":"Datos permitidos para la actualización de un plato existente (únicamente precio y descripción)"},"RestaurantRequestDto":{"required":["direccion","idPropietario","nit","nombre","telefono","urlLogo"],"type":"object","properties":{"nombre":{"type":"string","description":"Nombre comercial del restaurante (no puede estar compuesto únicamente por números)","example":"Restaurante El Buen Sabor"},"nit":{"type":"string","description":"Número de Identificación Tributaria (NIT únicamente numérico)","example":"900123456"},"direccion":{"type":"string","description":"Dirección física del restaurante","example":"Calle 100 # 15-20"},"telefono":{"type":"string","description":"Número telefónico de contacto (máximo 13 caracteres, admite prefijo +)","example":"+573001234567"},"urlLogo":{"type":"string","description":"URL accesible que apunta a la imagen del logo del restaurante","example":"https://misimagenes.com/logo-restaurante.png"},"idPropietario":{"type":"integer","description":"ID del usuario que actúa como propietario (debe existir y tener rol PROPIETARIO)","format":"int64","example":1}},"description":"Datos requeridos para la creación de un nuevo restaurante"},"DishRequestDto":{"required":["descripcion","idCategoria","idRestaurante","nombre","precio","urlImagen"],"type":"object","properties":{"nombre":{"type":"string","description":"Nombre comercial del plato","example":"Bandeja Paisa"},"idCategoria":{"type":"integer","description":"Identificador de la categoría a la que pertenece el plato","format":"int64","example":2},"descripcion":{"type":"string","description":"Descripción detallada de ingredientes y preparación del plato","example":"Deliciosa bandeja con frijoles, arroz, carne molida, chicharrón, huevo y aguacate"},"precio":{"minimum":1,"type":"integer","description":"Precio del plato (número entero positivo mayor a 0)","format":"int32","example":35000},"idRestaurante":{"type":"integer","description":"Identificador del restaurante al cual se asocia el plato","format":"int64","example":1},"urlImagen":{"type":"string","description":"URL accesible que apunta a la fotografía del plato","example":"https://misimagenes.com/platos/bandeja-paisa.png"}},"description":"Datos requeridos para la creación de un nuevo plato en el menú"}}}}